+++ src/java/growthcraft/apples/ClientProxy.java
@@ -1,5 +1,7 @@
 package growthcraft.apples;
 
+import growthcraft.apples.renderer.RenderApple;
+
 import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class ClientProxy extends CommonProxy
