+++ src/java/growthcraft/apples/block/BlockAppleLeaves.java
@@ -1,8 +1,12 @@
-package growthcraft.apples;
+package growthcraft.apples.block;
 
 import java.util.ArrayList;
 import java.util.Random;
 
+import growthcraft.apples.GrowthCraftApples;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockLeavesBase;
 import net.minecraft.block.material.Material;
@@ -16,8 +20,6 @@ import net.minecraft.world.ColorizerFoliage;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.IShearable;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockAppleLeaves extends BlockLeavesBase implements IShearable
 {
@@ -29,7 +31,7 @@ public class BlockAppleLeaves extends BlockLeavesBase implements IShearable
 	//public static Boolean graphicFlag;
 	int[] adjacentTreeBlocks;
 
-	protected BlockAppleLeaves()
+	public BlockAppleLeaves()
 	{
 		super(Material.leaves, false);
 		this.setTickRandomly(true);
