+++ src/java/growthcraft/apples/item/ItemAppleSeeds.java
@@ -1,6 +1,10 @@
-package growthcraft.apples;
+package growthcraft.apples.item;
 
+import growthcraft.apples.GrowthCraftApples;
 import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.Entity;
@@ -12,8 +16,6 @@ import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.EnumPlantType;
 import net.minecraftforge.common.IPlantable;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemAppleSeeds extends Item implements IPlantable
 {
